search: new anytime monitoring.
This is a simpler monitor that just returns whether or
not to stop, and not when to next monitor.  The monitoring
is cheap so we might as well do it all of the time.

Another benefit of this implementation is that it comes
with tools for plotting the policy (search/anyprof/sptpolicy),
and thus I am more confident that it is coming up with
policies that make sense.