gridnav: add a random instance generator.
This generator takes a map file as input and
simply generates two points in the map that are
no blocked.  The map followed by these points
are dumped to standard output.  This can be
used to make random instances from the
hog2 maps.  There is no guarantee that the
instances are solvable unless the maps are
fully connected.

There is a -s argument to specify a seed for the
random generator.  An instance is uniquely
defined by its map and seed.