plat2d: make the heuristic admissible!
All previous results were run with an inadmissible heuristic. It
didn't correctly account for the fact that you can move in both X and
Y directions at the same time, and sometimes would overestimate. The
new heuristic works as follows: build the visibility graph, however,
the cost of an edge is the max of the dx, dy, or diag/sqrt(2) of the
edge. This is admissible.